[contrib:elasticsearch] use class_eval for ES monkey patching (ruby19 compatible)
